feat (*) : delay line changed
Delay line changed to process each sample
seperately, so delay time can be changed with LFO.
Does not change the functionality of delay effect
but allows to add flanger, vibrato or chorus to class.
Also when delay time is changed, delay buffer is cleared
so there is no crackling from the change.